PCI: Use __pci_set_master from do_pci_disable_device
__pci_set_master is there to enable or disable the Bus Master Enable
bit in the COMMAND register.  This function also has debug log which is
useful to users.

do_pci_disable_device is also disabling the BME bitfield where some of
the codes are duplicated with __pci_set_master.  The only difference
between the two is whether to set flag `dev->is_busmaster` or not.

This patch removed that duplications by adding one more parameter to
__pci_set_master `update_bm` to decide whether to update the flag or
not.  It also will have the debug log from __pci_set_master which is
good for debug from logs.

Signed-off-by: Minwoo Im <minwoo.im.dev@gmail.com>