nvme: introduce passthrough ioctl for multipath
We don't allow NVME_IOCTL_IO_CMD ioctl in case that a controller has
multiple namespaces attached.  Also, I/O request to the controller
character device has not been recommended and deprecated because we have
block device to I/O with where the multipath consideration is taken.

Once kernel decided a path to I/O for a namespace based on the I/O
policy of a NVMe subsystem, userspace is not allowed to choose a path to
I/O.  If a path is broken(inaccessible state in ANA), then it will not
try to I/O to that path.

This patch introduced NVME_IOCTL_MPATH_IO command for controller
device(e.g., /dev/nvme0) to support multipath I/O passthrough for
userspace.  Regardless driver's path decision, userspace can target a
namespace to I/O.  In this case, `cmd.nsid` will be used to find out the
namespace instance target which is hidden(e.g., nvmeXcYnZ).

Signed-off-by: Minwoo Im <minwoo.im.dev@gmail.com>